+++ Mobile/src/reducers/auth_reducer.js
@@ -1,5 +1,8 @@
 'use strict';
 
+// Import libraries
+import axios from 'axios';
+
 // Import action types
 import {
 	EMAIL_CHANGED,
@@ -11,8 +14,11 @@ import {
 	SIGNUP_USER_FAIL,
 	LOGIN_USER,
 	LOGIN_USER_SUCCESS,
-	LOGIN_USER_FAIL
+	LOGIN_USER_FAIL,
+	SAVE_PROJECT_CHANGES
 } from '../actions/types';
+
+import { ROOT_URL } from '../constants';
 
 const INITIAL_STATE = {
 	email: '',
@@ -52,7 +58,23 @@ export default (state = INITIAL_STATE, action) => {
 			return { ...state, ...CLEAR_STATE, token: action.payload };
 		case LOGIN_USER_FAIL:
 			return { ...state, error: action.payload, password: '', loading: false };
+		case SAVE_PROJECT_CHANGES:
+			console.log("reducer fired in main");
+			addSubscriber(state.email, action.payload.id, action.change_type);
+			return state;
 		default:
 			return state;
 	}
+};
+
+const addSubscriber = (email, project_id, type) => {
+	console.log('POST request initiated to /addSubscriber');
+	console.log('body is: ', email, project_id, type);
+	axios.post(`${ROOT_URL}/addSubscriber`, {email, project_id, type})
+	.then((res) => {
+		console.log('addSubscriber successful. Response: ', res);
+	})
+	.catch((err) => {
+		console.log('addSubscriber FAIL. Response: ', err);
+	});
 };
